requirements.txt
langchain
langchain_community
langchain_openai
langgraph
youtube-transcript-api
fastapi
uvicorn
redis
elasticsearch
python-dotenv